+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased:
+### Changed:
+- Disable hardrules that need lang parameter when using multilingual model.
+
 ## Bicleaner AI 2.3.2:
 ### Fixed:
 - Fix compatibility with Hardrules metadata dict.